Change function views to class based views.

+++ multiupload/views.py
@@ -3,20 +3,23 @@
                 Sandeep Jadoonanan
 """
 
-from django.shortcuts import render
-from django.http import HttpResponse, Http404, HttpResponseBadRequest
-from django.conf import settings
 from django.shortcuts import get_object_or_404
-import os, json
-import uuid
+from django.http import HttpResponse, Http404
+from django.views.generic import View
+import json
 
 from .models import Image
 
 
-def handleUpload(request):
-  # This view accepts only POST methods.
-  if request.method == "POST":
-    
+class HandleUpload(View):
+  """
+  Handles photo uploads via the jQuery-file-upload plugin.
+  """
+  
+  def _allowed_methods(self):
+    return ["post"]
+   
+  def post(self, request):
     # Get the uploaded file.
     f = request.FILES["thefiles"]
     # Replace any whitespaces with underscores.
@@ -28,63 +31,65 @@ def handleUpload(request):
     
     # Create a JSON response (for Jquery File Upload).
     res = {
-      "files": []
+      "files": [imgFile.toObj()]
     }
-    res["files"].append(imgFile.toObj())
-    
     jsonResponse = json.dumps(res)
     
     return HttpResponse(jsonResponse, content_type = "application/json")
+
+
+class HandleDelete(View):
+  """
+  Deletes an image based on its hash value (imgHash).
+  """
+  
+  def _allowed_methods(self):
+    return ["post"]
+  
+  def post(self, request, *args, **kwargs):
+    # Get the image hash from url (Return 0 if no hash is found).
+    imageHash = kwargs.pop("hash", 0)
     
-  else:
-    return HttpResponseBadRequest("Only POST requests are allowed!", content_type = "application/json")
-   
-  raise Http404
+    if imageHash == 0:
+      raise Http404
+    
+    # Get Image object with the exact ImgHash as the url paramater 'hash'.
+    imgFile = get_object_or_404(Image, imgHash = imageHash)
+    
+    # Create a JSON response (for Jquery File Upload).
+    res = {}
+    res[imgFile.name] = True
+    res = {
+      "files": [res]
+    }
+    
+    # Delete the image
+    imgFile.delete()
+    jsonResponse = json.dumps(res)
+    
+    return HttpResponse(jsonResponse, content_type = "application/json")
 
 
-def handleDelete(request, hash):
-  # Get Image object with the exact ImgHash as the paramater 'hash'.
-  imgFile = get_object_or_404(Image, imgHash = hash)
-  
-  # Create a JSON response (for Jquery File Upload).
-  res = {}
-  res[imgFile.name] = True
-  res = {
-    "files": [res]
-  }
-  
-  # Delete the image
-  imgFile.delete()
-  jsonResponse = json.dumps(res)
-  
-  return HttpResponse(jsonResponse, content_type = "application/json")
-
-
-
-# NOTE: Make this ONLY POST requests.
-def returnImagesJSON(request):
+class ImagesJSON(View):
   """
   Return JSON data for all objects loaded. This way,
   AngularJS can construct the template on the front-end for us.
   """
-  # This view accepts only POST methods.
-  if request.method == "POST":
-     
+  
+  def _allowed_methods(self):
+    return ["post"]
+  
+  def post(self, request, *args, **kwargs):
+    # Limit the results to 'x' amount of Images.
     limit = 12
      
     ImageData = {
       "files": []
     }
-    # Limit the results to 'x' amount of Images.
+    
     for img in Image.objects.all()[:limit]:
       ImageData["files"].append(img.toObj(True))
      
     jsonResponse = json.dumps(ImageData)
      
     return HttpResponse(jsonResponse, content_type = "application/json")
-   
-  else:
-    return HttpResponseBadRequest("Only POST requests are allowed!", content_type = "application/json")
-   
-  raise Http404
-
